+++ src/libs/components/modal/modal-config.ts
@@ -3,8 +3,9 @@ import { ComponentType } from '@angular/cdk/portal';
 
 export class NtModalConfig<T = any> {
   id?: string;
-  width?: string = '400px';
-  height?: string = '600px';
+  // width?: string = '400px';
+  // height?: string = '600px';
+  top?: string = '80px';
   minWidth?: number | string;
   minHeight?: number | string;
   maxWidth?: number | string = '80vw';
@@ -16,4 +17,5 @@ export class NtModalConfig<T = any> {
   panelClass?: string = 'tm-file-preview-modal-panel';
   hasBackdrop?: boolean = true;
   backdropClass?: string = 'dark-backdrop';
+  centerVertically?: boolean = false;
 }